+++ libcloudvagrant/__init__.py
@@ -24,6 +24,8 @@ manage VirtualBox nodes, networks and volumes.
 """
 
 import os
+import re
+import subprocess
 import sys
 
 from libcloud.compute import providers as compute_providers
@@ -31,10 +33,10 @@ from libcloud.networking import providers as networking_providers
 from libcloud import security
 
 try:
-    import nose
-    NOSE_FOUND = True
+    import pytest
+    PYTEST_FOUND = True
 except ImportError:
-    NOSE_FOUND = False
+    PYTEST_FOUND = False
 
 try:
     import netifaces
@@ -60,7 +62,6 @@ __all__ = [
 compute_providers.set_driver(VAGRANT,
                              compute_driver.__name__,
                              compute_driver.VagrantDriver.__name__)
-
 networking_providers.set_driver(VAGRANT,
                                 networking_driver.__name__,
                                 networking_driver.VagrantDriver.__name__)
@@ -73,7 +74,7 @@ security.CA_CERTS_PATH.append(os.path.join(os.path.dirname(__file__),
 __version__ = "0.4.0"
 
 
-def test():
+def test():  # pragma: no cover
     """Runs the tests for ``libcloud-vagrant``.
 
     Requires ``nose`` to be installed. The tests take quite a bit of time to
@@ -82,8 +83,8 @@ def test():
 
     """
     can_run = True
-    if not NOSE_FOUND:
-        print >> sys.stderr, "Install 'nose' in order to run the tests."
+    if not PYTEST_FOUND:
+        print >> sys.stderr, "Install 'pytest' in order to run the tests."
         can_run = False
 
     if not NETIFACES_FOUND:
@@ -92,4 +93,44 @@ def test():
 
     if can_run:
         print "Testing libcloud-vagrant %s" % (__version__,)
-        return nose.run(__name__)
+        return pytest.main([__name__])
+
+
+def execute(cmdline):
+    p = subprocess.Popen(cmdline,
+                         shell=True,
+                         stdout=subprocess.PIPE,
+                         stderr=subprocess.STDOUT)
+    stdout, _ = p.communicate()
+    if p.returncode:
+        raise RuntimeError("Cannot execute '%s': %s", cmdline, stdout)
+    return stdout.strip()
+
+
+def check_versions():
+    """Verifies the versions of VirtualBox, Vagrant and the required Vagrant
+    plugins.
+
+    """
+    vagrant_version = execute("vagrant --version")
+    if not vagrant_version.startswith("Vagrant 1.6"):
+        raise RuntimeError("Unsupported %s" % (vagrant_version,))
+
+    virtualbox_version = execute("VBoxManage --version")
+    if not virtualbox_version.startswith("4.3"):
+        raise RuntimeError("Unsupported VirtualBox %s" %
+                           (virtualbox_version,))
+
+    vagrant_plugins = execute("vagrant plugin list")
+    m = re.search(r"vagrant-libcloud-helper \((.+?)\)", vagrant_plugins)
+    if not m:
+        print >> sys.stderr, "Installing plugin 'vagrant-libcloud-helper'"
+        execute("vagrant plugin update vagrant-libcloud-helper")
+    else:
+        required = "0.0.2"
+        if m.group(1) != required:
+            print >> sys.stderr, "Updating plugin 'vagrant-libcloud-helper'"
+            execute("vagrant plugin update vagrant-libcloud-helper")
+
+
+check_versions()