+++ libcloudvagrant/tests/test_network.py
@@ -25,9 +25,11 @@ import subprocess
 
 import netifaces
 
+from pytest import raises
+
 from libcloud.common.types import LibcloudError
 
-from libcloudvagrant.tests import new_driver, sample_network, sample_node
+from libcloudvagrant.tests import sample_network, sample_node
 
 
 __all__ = [
@@ -43,97 +45,84 @@ __all__ = [
 ]
 
 
-driver = new_driver()
-
-
-def test_create_network():
+def test_create_network(driver, network):
     """Network objects are opaque. For unit tests, some operations are
     supported.
 
     """
-    with sample_network("net1", cidr="192.168.0.0/24") as n:
-        assert len(n.addresses) == 254
-        for addr in n.addresses:
-            assert addr in n
+    assert len(network.addresses) == 254
+    for addr in network.addresses:
+        assert addr in network
 
 
-def test_create_duplicate_netwoks():
+def test_create_duplicate_netwoks(driver, network):
     """Networks with the same name may be created, as long as they are equal
     to those already created.
 
     """
-    with sample_network("net1", cidr="192.168.0.0/24") as net1:
-        net2 = driver.ex_create_network(name="net1", cidr="192.168.0.0/24")
-        assert net1 == net2
+    net2 = driver.ex_create_network(name=network.name, cidr=network.cidr)
+    assert net2 == network
 
-    with sample_network("net1", cidr="192.168.0.0/24") as net1:
-        try:
-            driver.ex_create_network(name="net1", cidr="10.0.0.0/8")
-        except Exception, exc:
-            assert exc.value == "Network 'net1' already defined"
-        else:
-            raise AssertionError("Non-equal duplicate networks "
-                                 "cannot be created")
+    other_cidr = "192.168.0.0/24"
+    with raises(LibcloudError) as exc:
+        driver.ex_create_network(name=network.name, cidr=other_cidr)
+    assert exc.value.value == ("Network '%s' already defined" %
+                               (network.name,))
 
 
-def test_destroy_network():
+def test_destroy_network(driver, network):
     """Networks in use may not be destroyed.
 
     """
-    with sample_network("net1") as net1:
-        with sample_node(networks=[net1]):
-            assert not driver.ex_destroy_network(net1)
+    with sample_node(driver, networks=[network]):
+        assert not driver.ex_destroy_network(network)
 
-        assert driver.ex_destroy_network(net1)
+    assert driver.ex_destroy_network(network)
 
 
-def test_exhausted_networks():
+def test_exhausted_networks(driver):
     """Networks cannot be over-allocated.
 
     """
-    with sample_network("net1", cidr="172.16.0.0/30", public=True) as net:
+    with sample_network(driver, "net1", cidr="172.16.0.0/30", public=True) as net:
         addr = net.allocate_address()
         assert str(addr.address) == "172.16.0.2"
 
-        try:
+        with raises(LibcloudError) as exc:
             addr = net.allocate_address()
-        except LibcloudError, exc:
-            assert exc.value == "No more free addresses in 172.16.0.0/30"
-        else:
-            raise AssertionError("Address %s should not be available" %
-                                 (addr,))
+        assert exc.value.value == "No more free addresses in 172.16.0.0/30"
 
 
-def test_host_interface_cleanup():
+def test_host_interface_cleanup(driver):
     """Host interfaces are removed when their associated networks are
     destroyed.
 
     """
     assert not any(i.startswith("vboxnet") for i in netifaces.interfaces())
 
-    with sample_network("pub", public=True) as pub:
-        with sample_node(networks=[pub]):
+    with sample_network(driver, "pub", public=True) as pub:
+        with sample_node(driver, networks=[pub]):
             iface = pub.host_interface
             assert iface in netifaces.interfaces()
         assert iface in netifaces.interfaces()
     assert iface not in netifaces.interfaces()
 
-    with sample_network("priv") as priv:
-        with sample_node(networks=[priv]):
+    with sample_network(driver, "priv") as priv:
+        with sample_node(driver, networks=[priv]):
             assert priv.host_interface is None
 
 
-def test_list_networks():
+def test_list_networks(driver):
     """Networks are listed properly.
 
     """
     assert len(driver.ex_list_networks()) == 0
-    with sample_network("net1") as net1:
+    with sample_network(driver, "net1") as net1:
         networks = driver.ex_list_networks()
         assert len(networks) == 1
         assert networks[0] == net1
 
-        with sample_network("net2") as net2:
+        with sample_network(driver, "net2") as net2:
             networks = driver.ex_list_networks()
             assert len(networks) == 2
             assert net1 in networks
@@ -146,72 +135,41 @@ def test_list_networks():
     assert len(driver.ex_list_networks()) == 0
 
 
-def test_overlapping_networks():
+def test_overlapping_networks(driver):
     """Overlapping networks are not supported.
 
     """
-    with sample_network("net1", cidr="192.168.0.0/24"):
-        try:
+    with sample_network(driver, "net1", cidr="192.168.0.0/24"):
+        with raises(LibcloudError) as exc:
             driver.ex_create_network(name="net2",
                                      cidr="192.168.0.0/23")
-        except Exception, exc:
-            assert exc.value == "Network 'net2' overlaps with 'net1'"
-        else:
-            raise AssertionError("Overlapping networks are not supported")
+        assert exc.value.value == "Network 'net2' overlaps with 'net1'"
 
-        try:
+        with raises(LibcloudError) as exc:
             driver.ex_create_network(name="net2",
                                      cidr="192.168.0.0/25")
-        except Exception, exc:
-            assert exc.value == "Network 'net2' overlaps with 'net1'"
-        else:
-            raise AssertionError("Overlapping networks are not supported")
+        assert exc.value.value == "Network 'net2' overlaps with 'net1'"
 
 
-def test_public_network():
-    with sample_network("public", public=True) as pub:
-        with sample_node(networks=[pub]) as n:
-            assert len(n.public_ips) == 1
-            addr = n.public_ips[0]
-            assert addr != pub.host_address
-            assert ping(addr)
-            assert not n.private_ips
+def test_public_network(driver, public_network):
+    with sample_node(driver, networks=[public_network]) as n:
+        assert len(n.public_ips) == 1
+        addr = n.public_ips[0]
+        assert addr != public_network.host_address
+        assert ping(addr)
+        assert not n.private_ips
 
 
-def test_public_and_private_networks():
+def test_public_and_private_networks(driver, public_network, private_network):
     """Public networks are accessible (within the host), but private ones are
     not.
 
     """
-    pub = sample_network("public", public=True)
-    dmz = sample_network("dmz")
-    lan = sample_network("lan")
-
-    with pub as pub, dmz as dmz, lan as lan:
-
-        nginx = sample_node(networks=[pub, dmz])
-        django = sample_node(networks=[dmz, lan])
-        postgresql = sample_node(networks=[lan])
-
-        with nginx as nginx, django as django, postgresql as postgresql:
-            assert len(nginx.public_ips) == 1
-            addr = nginx.public_ips[0]
-            assert addr != pub.host_address
-            assert ping(addr)
-
-            assert len(nginx.private_ips) == 1
-            addr = nginx.private_ips[0]
-            assert not ping(addr)
-
-            assert len(django.public_ips) == 0
-            assert len(django.private_ips) == 2
-            for addr in django.private_ips:
-                assert not ping(addr)
-
-            assert len(postgresql.public_ips) == 0
-            assert len(postgresql.private_ips) == 1
-            addr = postgresql.private_ips[0]
-            assert not ping(addr)
+    with sample_node(driver, networks=[public_network, private_network]) as n:
+        assert len(n.public_ips) == 1
+        assert ping(n.public_ips[0])
+        assert len(n.private_ips) == 1
+        assert not ping(n.private_ips[0])
 
 
 LOG = logging.getLogger("libcloudvagrant")