+++ libcloudvagrant/tests/test_types.py
@@ -20,14 +20,15 @@
 
 """Unit tests for data types."""
 
-from libcloudvagrant.tests import new_driver
-from libcloudvagrant.common.types import (
-    VagrantAddress,
+from libcloudvagrant.compute.types import (
     VagrantImage,
-    VagrantNetwork,
     VagrantNode,
     VagrantNodeSize,
     VagrantVolume,
+)
+from libcloudvagrant.networking.types import (
+    VagrantAddress,
+    VagrantNetwork,
 )
 
 
@@ -35,9 +36,6 @@ __all__ = [
     "test_serializable",
     "test_serializable_with_driver",
 ]
-
-
-driver = new_driver()
 
 
 def test_serializable():
@@ -67,16 +65,18 @@ def test_serializable():
                         })
 
 
-def test_serializable_with_driver():
+def test_serializable_with_driver(driver):
     """Serializable types can be converted to and from their dict
     representations.
 
     """
-    assert_serializable_with_driver(VagrantImage,
+    assert_serializable_with_driver(driver,
+                                    VagrantImage,
                                     {
                                         "name": "ubuntu/trusty64",
                                     })
-    assert_serializable_with_driver(VagrantNode,
+    assert_serializable_with_driver(driver,
+                                    VagrantNode,
                                     {
                                         "id": "27036b03-13a1-45d6-9030-a3faef699ba9",
                                         "name": "node1",
@@ -88,6 +88,7 @@ def test_serializable_with_driver():
                                             "cpus": 3,
                                         },
                                         "image": {"name": "ubuntu/trusty64"},
+                                        "allocate_sata_ports": 30,
                                     },
                                     {
                                         "id": "361f2550-bc23-4eca-ad22-49914dd8b530",
@@ -110,14 +111,17 @@ def test_serializable_with_driver():
                                             "cpus": 2,
                                         },
                                         "image": {"name": "ubuntu/trusty64"},
+                                        "allocate_sata_ports": 30,
                                     })
-    assert_serializable_with_driver(VagrantNodeSize,
+    assert_serializable_with_driver(driver,
+                                    VagrantNodeSize,
                                     {
                                         "name": "default",
                                         "ram": 2048,
                                         "cpus": 2,
                                     })
-    assert_serializable_with_driver(VagrantVolume,
+    assert_serializable_with_driver(driver,
+                                    VagrantVolume,
                                     {
                                         "name": "test-volume",
                                         "size": 42,
@@ -145,7 +149,7 @@ def assert_serializable(cls, *params):
         assert recreated == obj
 
 
-def assert_serializable_with_driver(cls, *params):
+def assert_serializable_with_driver(driver, cls, *params):
     """Serializable types can be converted to and from their dict
     representations.